requirements.txt
Django==5.0.1
django-rq==2.10.2
anthropic==0.25.1
Pillow==10.2.0
python-decouple==3.8
django-cors-headers==4.3.1
redis==5.0.1
rq==1.15.1